% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define}
\usage{
define(..., ignore = FALSE, input = NULL, output = NULL, type = NULL)
}
\arguments{
\item{...}{Named arguments to define.}

\item{ignore}{Whether to ignore the arguments from fields and function call.}
}
\description{
Define a function or argument for processing.
}
