% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define}
\usage{
define(
  ...,
  ignore = FALSE,
  registry_input = NULL,
  registry_output = NULL,
  type = NULL,
  title = NULL,
  registry_name = NULL,
  registry_description = NULL,
  description = NULL,
  classes = NULL,
  output_function = NULL,
  render_function = NULL,
  evaluate_function = NULL,
  block_combiner = NULL,
  generate_server_function = NULL,
  download_ui = NULL,
  download_server = NULL
)
}
\arguments{
\item{...}{Named arguments to define.}

\item{ignore}{Whether to ignore the arguments from fields and function call.}

\item{registry_input, registry_output}{Object accepted as I/O to generate the block registration.}

\item{type}{Block type, used for inheritance.}

\item{title, description}{Title and description of field.}

\item{registry_name, registry_description}{Name and description in of the block in the registry.}

\item{classes}{Additional classes.}

\item{output_function}{Function to generate output method.}

\item{render_function}{Function to generate render method.}

\item{evaluate_function}{Function to generate evaluate method.}

\item{block_combiner}{Function to combine blocks.}

\item{generate_server_function}{Function to generate server function.}

\item{download_ui, download_server}{Download UI and Server methods.}
}
\description{
Define a function or argument for processing.
}
