% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_blocks}
\alias{generate_blocks}
\alias{generate_block}
\alias{generate_block.function}
\alias{generate_block.character}
\title{Generate Blocks}
\usage{
generate_blocks(
  package,
  filter_functions = identity,
  functions = define(),
  all_functions = define()
)

generate_block(
  fn,
  filter_functions = identity,
  functions = define(),
  all_functions = define()
)

\method{generate_block}{`function`}(
  fn,
  filter_functions = identity,
  functions = define(),
  all_functions = define()
)

\method{generate_block}{character}(
  fn,
  filter_functions = identity,
  functions = define(),
  all_functions = define()
)
}
\arguments{
\item{package}{Name of package.}

\item{filter_functions}{callback that filters functions,
it must accept a character vector and return a character vector.}

\item{functions, all_functions}{Functions to define.}

\item{fn}{Function to create a block for.}
}
\description{
Generate block code from package in \verb{R/} directory.
}
