% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{generate_blocks}
\alias{generate_blocks}
\title{Generate Blocks}
\usage{
generate_blocks(
  package,
  default_type = c("transform", "data", "plot"),
  default_input = "data.frame",
  default_output = "data.frame",
  ignore_args = c(),
  ...
)
}
\arguments{
\item{package}{Name of package.}

\item{default_type}{Default block type.}

\item{default_input}{Default block input type.}

\item{default_output}{Default block output type.}

\item{ignore_args}{Arguments to omit from fields.}

\item{...}{Named list of special fields for each function.}
}
\description{
Generate block code from package in \verb{R/} directory.
}
