% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_blocks}
\alias{generate_blocks}
\title{Generate Blocks}
\usage{
generate_blocks(
  package,
  filter_functions = identity,
  class = NULL,
  functions = define(),
  all_functions = define()
)
}
\arguments{
\item{package}{Name of package.}

\item{filter_functions}{callback that filters functions,
it must accept a character vector and return a character vector.}

\item{class}{Additional class to pass to block.}

\item{functions, all_functions}{Functions to define.}
}
\description{
Generate block code from package in \verb{R/} directory.
}
