% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_blocks}
\alias{generate_blocks}
\title{Generate Blocks}
\usage{
generate_blocks(
  package,
  default_type = c("transform", "data", "plot"),
  default_input = "data.frame",
  default_output = "data.frame",
  class = NULL,
  config = define(),
  all_args = define()
)
}
\arguments{
\item{package}{Name of package.}

\item{default_type}{Default block type.}

\item{default_input}{Default block input type.}

\item{default_output}{Default block output type.}

\item{class}{Additional class to pass to block.}

\item{config, all_args}{Configuration for blocks.}
}
\description{
Generate block code from package in \verb{R/} directory.
}
