% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_blocks}
\alias{generate_blocks}
\title{Generate Blocks}
\usage{
generate_blocks(
  package,
  class = NULL,
  functions = define(),
  all_functions = define()
)
}
\arguments{
\item{package}{Name of package.}

\item{class}{Additional class to pass to block.}

\item{functions, all_functions}{Functions to define.}

\item{default_input}{Default block input type.}

\item{default_output}{Default block output type.}
}
\description{
Generate block code from package in \verb{R/} directory.
}
